Move parser and model out to their own file
